+++ CMakeLists.txt
@@ -6,7 +6,6 @@ set(CMAKE_CXX_STANDARD 17)
 
 find_package(Boost 1.66.0 REQUIRED COMPONENTS system filesystem)
 include_directories (${Boost_INCLUDE_DIR})
-#find_package(nlohmann_json REQUIRED)
 
 add_subdirectory(src)
